--- mysql.js
@@ -1,19 +1,0 @@
-var mysql = require('mysql');
-
-var db = mysql.createConnection({
-    host     : 'localhost',
-    user     : 'root',
-    password : '123456',
-    database : 'opentt'
-});
-
-db.connect();
-
-db.query('SELECT * FROM topic', function (error, results, fields) {
-    if (error) {
-        console.log(error);
-    }
-    console.log(results);
-});
-
-db.end();
